pyproject.toml
[project]
name = "planetclicker"
version = "0.0.1"
description = "Clicker game."
readme = "README.md"
requires-python = ">=3.12"
authors = [
  {name = "Ben Kanter"},
  {name = "Adam Neeley"}
]
dependencies = [
  "behave",
  "numpy",
  "pygame_ce",
  "pygame_gui",
  "pytest",
  "pytest-cov",
  "toml",
  "tomlkit",
]
[project.urls]
repository = "https://github.com/wwafflless/PlanetClicker"


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
pythonpath = "."
minversion = "6.0"
addopts = "-ra -q"
testpaths = ["tests"]

[tool.pyright]
pythonpath = ["planetclicker"]
include = ["planetclicker", "tests", "features"]
exclude = ["**/__pycache__"]
executionEnvironments = [
  { root = "." },
]

[project.scripts]
main = "planetclicker.__main__:main"
